Refactor region import and export into separate files in new dir
